+++ CHANGELOG.md
@@ -33,6 +33,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
   variable has no time variation.
 * Make the performance breakdown section of the model summary report display
   both ROI and CPIK charts for all scenarios.
+* Set default ROI priors for non-revenue, no revenue-per-KPI models.
 
 ## [0.7.0] - 2024-09-20